pyproject.toml
[project]
name = "vietlegal-e5"
version = "0.1.0"
description = "Fine-tune multilingual-e5-large for Vietnamese legal retrieval"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=3.0,<4.0",
    "transformers>=4.40,<5.0",
    "datasets>=2.18,<3.0",
    "torch>=2.1,<3.0",
    "accelerate>=0.28,<1.0",
    "faiss-cpu>=1.7,<2.0",
    "wandb>=0.16,<1.0",
    "pandas>=2.0,<3.0",
    "pyarrow>=18.0",
    "pyyaml>=6.0,<7.0",
    "rank_bm25>=0.2.2,<1.0",
    "optimum[onnxruntime]>=1.17,<2.0",
    "huggingface_hub>=0.21,<1.0",
    "openai>=1.12,<2.0",
    "scikit-learn>=1.3,<2.0",
]

[project.optional-dependencies]
vllm = ["vllm>=0.4,<1.0"]
